# CSharp/MotionController/MotionController Sensor API/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Internal MotionController Sensor API",
    "description": "API interface for all internal sensor operations.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/device/sessions/{sessionId}": {
      "get": {
        "tags": [
          "DeviceSession"
        ],
        "summary": "Get a Device Session by Session Id",
        "operationId": "GetDeviceSessionBySessionIdAsync",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSession"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/game/sessions/{sessionId}": {
      "get": {
        "tags": [
          "GameSession"
        ],
        "summary": "Get a Game Session by Session Id",
        "operationId": "GetGameSessionBySessionIdAsync",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSession"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/game/sessions": {
      "post": {
        "tags": [
          "GameSession"
        ],
        "summary": "Adds a Game Session",
        "operationId": "AddGameSessionAsync",
        "requestBody": {
          "x-name": "gameSession",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnityGameSession"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DatabaseModel": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false
      },
      "DeviceSession": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DatabaseModel"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "integer",
                "format": "int32"
              },
              "sessionId": {
                "type": "string",
                "format": "guid"
              },
              "created": {
                "type": "string",
                "format": "date-time"
              },
              "modified": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
...
      "UnityGameSession": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "guid": {
            "type": "string",
            "format": "guid"
          },
          "playerData": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PlayerData"
              }
            ]
          },
          "gameData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameData"
            }
          }
        }
      },
      "PlayerData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "lives": {
            "type": "integer",
            "format": "int32"
          },
          "gameTime": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GameData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ballPosition": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Vector3"
              }
            ]
          },
          "boardRotation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Vector3"
              }
            ]
          },
          "inputData": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Vector2"
              }
            ]
          }
        }
      },
      "Vector3": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          },
          "z": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "Vector2": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "x": {
            "type": "number",
            "format": "float"
          },
          "y": {
            "type": "number",
            "format": "float"
          }
        }
      }
    }
  }
}